chore(socket-validation): split message validators
